pc - dry up sample user code in storybook
