+++ stories/setup.js
@@ -1,0 +1,12 @@
+import { select, text } from "@storybook/addon-knobs";
+
+export const sampleUser = () => {
+  const name = text("Name", "Phill Conrad");
+  const role = select("Role", ["admin", "student", "guest"], "guest");
+  const picture = text(
+    "Image URL",
+    "https://avatars3.githubusercontent.com/u/1119017"
+  );
+  const user = { name, role, picture };
+  return user;
+};
